moving unbounded buffered chat to a different file
