+++ python/packages/autogen-core/src/autogen_core/model_context/__init__.py
@@ -1,9 +1,9 @@
-from ._buffered_chat_completion_context import (
-    BufferedChatCompletionContext,
-    UnboundedBufferedChatCompletionContext,
-)
+from ._buffered_chat_completion_context import BufferedChatCompletionContext
 from ._chat_completion_context import ChatCompletionContext
 from ._head_and_tail_chat_completion_context import HeadAndTailChatCompletionContext
+from ._unbounded_buffered_chat_completion_context import (
+    UnboundedBufferedChatCompletionContext,
+)
 
 __all__ = [
     "ChatCompletionContext",
