+++ python/packages/autogen-core/src/autogen_core/model_context/_buffered_chat_completion_context.py
@@ -43,36 +43,3 @@ class BufferedChatCompletionContext(ChatCompletionContext):
     def load_state(self, state: Mapping[str, Any]) -> None:
         self._messages = state["messages"]
         self._buffer_size = state["buffer_size"]
-
-
-class UnboundedBufferedChatCompletionContext(ChatCompletionContext):
-    """A buffered chat completion context that keeps a view of the last n messages,
-    where n is the buffer size. The buffer size is set at initialization.
-
-    Args:
-        buffer_size (int): The size of the buffer.
-
-    """
-
-    def __init__(self, initial_messages: List[LLMMessage] | None = None) -> None:
-        self._messages: List[LLMMessage] = initial_messages or []
-
-    async def add_message(self, message: LLMMessage) -> None:
-        """Add a message to the memory."""
-        self._messages.append(message)
-
-    async def get_messages(self) -> List[LLMMessage]:
-        """Get at most `buffer_size` recent messages."""
-        return self._messages
-
-    async def clear(self) -> None:
-        """Clear the message memory."""
-        self._messages = []
-
-    def save_state(self) -> Mapping[str, Any]:
-        return {
-            "messages": [message for message in self._messages],
-        }
-
-    def load_state(self, state: Mapping[str, Any]) -> None:
-        self._messages = state["messages"]
